+++ src/test/java/com/braintreegateway/unittest/CustomerRequestTest.java
@@ -15,19 +15,20 @@ public class CustomerRequestTest {
   @Test
   public void toXmlIncludesAllElements() throws IOException, SAXException {
     CustomerRequest request = new CustomerRequest()
-      .deviceData("{\"device_session_id\": \"devicesession123\", \"fraud_merchant_id\": \"fraudmerchant456\"}")
       .company("some-company")
       .customerId("customer-id")
+      .customField("some-custom-field", "some-custom-value")
+      .defaultPaymentMethodToken("some-token")
+      .deviceData("{\"device_session_id\": \"devicesession123\", \"fraud_merchant_id\": \"fraudmerchant456\"}")
       .email("some-email")
       .fax("some-fax-number")
       .firstName("Dan")
       .id("some-id")
       .lastName("Schulman")
+      .paymentMethodNonce("some-nonce")
       .phone("some-phone-number")
       .website("some-website")
-      .paymentMethodNonce("some-nonce")
-      .defaultPaymentMethodToken("some-token")
-      .customField("some-custom-field", "some-custom-value")
+      .androidPayCard().number("12345").done()
       .applePayCard().number("12345").done()
       .creditCard().number("12345").done()
       .riskData().customerIP("6789").done()
@@ -39,31 +40,33 @@ public class CustomerRequestTest {
 
     String expectedXML =
       "<customer>\n"
-      + "  <deviceData>{&quot;device_session_id&quot;: &quot;devicesession123&quot;, &quot;fraud_merchant_id&quot;: &quot;fraudmerchant456&quot;}</deviceData>\n"
+      + "  <androidPayCard>\n"
+      + "    <number>12345</number>\n"
+      + "  </androidPayCard>\n"
+      + "  <applePayCard>\n"
+      + "    <number>12345</number>\n"
+      + "  </applePayCard>\n"
       + "  <company>some-company</company>\n"
+      + "  <creditCard>\n"
+      + "    <number>12345</number>\n"
+      + "  </creditCard>\n"
+      + "  <customFields>\n"
+      + "    <some-custom-field>some-custom-value</some-custom-field>\n"
+      + "  </customFields>\n"
+      + "  <defaultPaymentMethodToken>some-token</defaultPaymentMethodToken>\n"
+      + "  <deviceData>{&quot;device_session_id&quot;: &quot;devicesession123&quot;, &quot;fraud_merchant_id&quot;: &quot;fraudmerchant456&quot;}</deviceData>\n"
       + "  <email>some-email</email>\n"
       + "  <fax>some-fax-number</fax>\n"
       + "  <firstName>Dan</firstName>\n"
       + "  <id>some-id</id>\n"
       + "  <lastName>Schulman</lastName>\n"
-      + "  <phone>some-phone-number</phone>\n"
-      + "  <website>some-website</website>\n"
-      + "  <paymentMethodNonce>some-nonce</paymentMethodNonce>\n"
-      + "  <defaultPaymentMethodToken>some-token</defaultPaymentMethodToken>\n"
-      + "  <applePayCard>\n"
-      + "    <number>12345</number>\n"
-      + "  </applePayCard>\n"
-      + "  <creditCard>\n"
-      + "    <number>12345</number>\n"
-      + "  </creditCard>\n"
       + "  <options>\n"
       + "    <paypal>\n"
       + "      <amount>10</amount>\n"
       + "    </paypal>\n"
       + "  </options>\n"
-      + "  <customFields>\n"
-      + "    <some-custom-field>some-custom-value</some-custom-field>\n"
-      + "  </customFields>\n"
+      + "  <paymentMethodNonce>some-nonce</paymentMethodNonce>\n"
+      + "  <phone>some-phone-number</phone>\n"
       + "  <riskData>\n"
       + "    <customerIP>6789</customerIP>\n"
       + "  </riskData>\n"
@@ -73,6 +76,7 @@ public class CustomerRequestTest {
       + "      <identifier>123</identifier>\n"
       + "    </taxIdentifier>\n"
       + "  </taxIdentifiers>\n"
+      + "  <website>some-website</website>\n"
       + "</customer>";
 
     XMLUnit.setIgnoreWhitespace(true);